Remove slash requirement in site.baseurl
Out of the box, in order for jekyll_uno to work, a trailing slash was
required in the baseurl inside of _config.yml. When no base url is used,
this caused multiple built pages to contain duplicade slashes. Adjusted
all layouts, includes, index.html, and contact.html to allow for a blank
site.baseurl in config.